implemented bidder offers list
