--- src/components/BidDetails.vue
@@ -1,554 +1,0 @@
-<template>
-  <v-toolbar color="blue" extended extension-height="100">
-    <v-chip :to="{ name: 'tenders' }" style="margin-left: 12rem" variant="text" text-color="white"
-      prepend-icon="mdi-keyboard-backspace">Back
-    </v-chip>
-    <template v-slot:extension>
-      <v-container class="px-15">
-        <v-toolbar-title class="ml-14 mb-4" style="font-size: 1.5rem">{{ tender.cpv.summary }} ({{ tender.cpv.code }})</v-toolbar-title>
-          <div style="margin-left: 3rem; margin-bottom: 3rem;">
-            <div v-if="$route.params.role === USER_ROLE.CONTRACTOR">
-              <div v-if="isTenderDecription">
-                <v-btn v-if="Number($route.query.offers) > 0" @click="navigateToOffers" rounded="0">Offers</v-btn>
-                <v-btn @click="tab='tenderDescription'" rounded="0">Tender Description</v-btn>
-              </div>
-            </div>
-          <div v-if="$route.params.role === USER_ROLE.BIDDER && Number($route.query.offer_id) > 0">
-            <v-btn v-if="offer.hasAwardDecision" @click="tab='awardDecision'" rounded="0">Award Decision</v-btn>
-            <v-btn v-if="offer.hasRejectDecision" @click="tab='rejectDecision'" rounded="0">Reject Decision</v-btn>
-            <v-btn @click="tab='tenderDescription'" rounded="0">Tender Description</v-btn>
-            <v-btn @click="navigateToOffer($route.query.offer_id)" rounded="0">My Offer</v-btn>
-          </div>
-        </div>
-      </v-container>
-    </template>
-  </v-toolbar>
-
-  <v-window v-model="tab" class="mt-n8 pb-10">
-    <v-window-item value="offers">
-      <v-card elevation="8" width="1000" class="mx-auto">
-       <TableHeaderWithoutField
-        companyName="Oficial Name"
-        price="Price"
-        country="Country"
-        date="Received Date"
-        status="Status"
-      ></TableHeaderWithoutField>
-      <v-container
-        id="scroll-target"
-        style="max-height: 25rem"
-        class="overflow-y-auto"
-        v-scroll:#scroll-target="onScroll">
-      <TableBody
-        :offers="offers"
-        @select-offer="(offer) => navigateToOffer(offer.id)"
-      ></TableBody>
-      </v-container>
-      </v-card>
-    </v-window-item>
-
-    <v-window-item value="awardDecision">
-      <v-card class="mx-auto" elevation="8" max-width="1000">
-        <v-toolbar color="white" height="280" class="text-left">
-          <v-toolbar-title v-if="OFFER_STATUS[offer.status] === OFFER_STATUS.OFFER_SELECTED_BY_CONTRACTOR" class="text-center" style="font-size: 1.5rem">
-            “Congratulation! Your Offer was selected by Contractor”
-          </v-toolbar-title>
-          <v-toolbar-title v-else-if="OFFER_STATUS[offer.status] === OFFER_STATUS.CONTRACT_APPROVED_BY_BIDDER" class="text-center" style="font-size: 1.5rem">
-            “Contract is approved by Bidder”
-          </v-toolbar-title>
-            <v-toolbar-title v-if="OFFER_STATUS[offer.status] === OFFER_STATUS.CONTRACT_DECLINED_BY_BIDDER" class="text-center" style="font-size: 1.5rem">
-            “Contract is declined by Bidder”
-          </v-toolbar-title>
-        </v-toolbar>
-          <v-item-group>
-        <v-row justify="center mb-10">
-          <v-col md = 4 class="text-center">
-          <FileVchip
-             :fileName="tender.awardDecision.fileMetadata.name"
-             :fileKey="tender.awardDecision.fileMetadata.awsS3fileKey"
-            @show-file="showFile"
-          ></FileVchip>
-        </v-col>
-          </v-row>
-      </v-item-group>
-        <v-row v-if="OFFER_STATUS[offer.status] === OFFER_STATUS.OFFER_SELECTED_BY_CONTRACTOR" class="d-flex justify-center mt-2 mb-10">
-      <v-col md="3">
-        <v-btn block variant="outlined" color="blue" @click="declineContract({ contractId: contract.id })">
-            Decline
-        </v-btn>
-         </v-col>
-          <v-col md="3">
-           <v-btn block variant="flat" color="blue"
-             @click="signContract({ contractId: contract.id,
-                                    offerId: offer.id,
-                                    rejectId: tender.rejectDecision.id })">
-             Approve
-           </v-btn>
-         </v-col>
-       </v-row>
-      </v-card>
-    </v-window-item>
-
-    <v-window-item value="rejectDecision">
-      <v-card class="mx-auto" elevation="8" max-width="1000">
-        <v-toolbar color="white" height="280" class="text-left">
-          <v-toolbar-title class="text-center" style="font-size: 1.5rem">
-            “Offer is rejected by contractor”
-          </v-toolbar-title>
-        </v-toolbar>
-        <v-item-group>
-        <v-row justify="center mb-10">
-          <v-col md = 4 class="text-center">
-          <FileVchip
-             :fileName="tender.rejectDecision.fileMetadata.name"
-             :fileKey="tender.rejectDecision.fileMetadata.awsS3fileKey"
-            @show-file="showFile"
-          ></FileVchip>
-        </v-col>
-          </v-row>
-      </v-item-group>
-      </v-card>
-    </v-window-item>
-
-    <v-window-item value="tenderDescription">
-      <v-card class="mx-auto" elevation="8" max-width="1000" >
-        <v-toolbar color="white" height="240" class="text-left">
-          <v-container class="pa-15">
-            <v-toolbar-title class="mb-5">Contractor</v-toolbar-title>
-            <v-row class="text-center mx-10">
-              <v-col class="text-left">
-                <div class="ma-2 details-title">Oficial Name:</div>
-                <div class="ma-2 details-title">National Registration Number: </div>
-              </v-col>
-              <v-col class="text-left">
-                <div class="ma-2">{{ tender.companyProfile.officialName }}</div>
-                <div class="ma-2">{{ tender.companyProfile.registrationNumber }}</div>
-              </v-col>
-              <v-col class="text-left">
-                <div class="ma-2 details-title">Country:</div>
-                <div class="ma-2 details-title">City / Town:</div>
-              </v-col>
-              <v-col class="text-left">
-                <div class="ma-2">{{ tender.companyProfile.country.name }}</div>
-                <div class="ma-2">{{ tender.companyProfile.city }}</div>
-              </v-col>
-            </v-row>
-          </v-container>
-        </v-toolbar>
-      </v-card>
-
-      <v-container class="d-flex align-center details-container">
-        <p class="chapter-label">Contact person</p>
-        <div class="horizontal-divider-details"></div>
-      </v-container>
-
-      <v-container class="details-container">
-        <v-row class="text-center mx-10">
-          <v-col class="text-left mx-2">
-            <div class="details-title">Name:</div>
-            <div class="mt-3 details-title">Surname:</div>
-          </v-col>
-          <v-col class="text-left mx-2">
-            <div>{{ tender.companyProfile.contactPerson.firstName }}</div>
-            <div class="mt-3">{{ tender.companyProfile.contactPerson.lastName }}</div>
-          </v-col>
-          <v-col class="text-left mx-2">
-            <div class="details-title">Phone number:</div>
-          </v-col>
-          <v-col class="text-left mx-2">
-            <div>{{ tender.companyProfile.contactPerson.phoneNumber }}</div>
-          </v-col>
-        </v-row>
-      </v-container>
-
-      <v-container class="d-flex align-center details-container">
-        <p class="chapter-label">Procedure</p>
-        <div class="horizontal-divider-details"></div>
-      </v-container>
-
-      <v-container class="details-container">
-        <v-row class="text-center mx-10">
-          <v-col class="text-left mx-2">
-            <div class="details-title">Procedure:</div>
-          </v-col>
-          <v-col class="text-left mx-2">
-            <div>{{ PROCEDURE[tender.procedure.type] }}</div>
-          </v-col>
-          <v-col class="text-left mx-2">
-            <div class="details-title">Language:</div>
-          </v-col>
-          <v-col class="text-left mx-2">
-            <div>{{ LANGUAGE[tender.procedure.language] }}</div>
-          </v-col>
-        </v-row>
-      </v-container>
-
-      <v-container class="d-flex align-center details-container">
-        <p class="chapter-label">Subject matter of the procurement</p>
-        <div class="horizontal-divider-details"></div>
-      </v-container>
-
-      <v-container class="details-container">
-        <v-row class="text-center mx-10">
-          <v-col class="text-left mx-2">
-            <div class="details-title">CPV Code:</div>
-            <div class="mt-3 details-title">Min Tender Value:</div>
-            <div class="mt-3 details-title">Currency:</div>
-          </v-col>
-          <v-col class="text-left mx-2">
-            <div>{{ tender.cpv.code }}</div>
-            <div class="mt-3">{{ contract.minPrice }}</div>
-            <div class="mt-3">{{ `${contract.currency.code} | ${contract.currency.symbol}` }}</div>
-          </v-col>
-          <v-col class="text-left mx-2">
-            <div class="details-title">Type of Tender:</div>
-            <div class="mt-3 details-title">Max Tender Value:</div>
-            <div class="mt-3 details-title">Description:</div>
-          </v-col>
-          <v-col class="text-left mx-2">
-            <div>{{ CONTRACT_TYPE[contract.contractType.title] }}</div>
-            <div class="mt-3">{{ contract.maxPrice }}</div>
-            <div class="mt-3">{{ tender.description }}</div>
-          </v-col>
-        </v-row>
-      </v-container>
-
-      <v-container class="d-flex align-center details-container">
-        <p class="chapter-label">Date</p>
-        <div class="horizontal-divider-details"></div>
-      </v-container>
-
-      <v-container class="details-container">
-        <v-row class="text-center mx-10">
-          <v-col class="text-left mx-2">
-            <div class="details-title">Publication Date:</div>
-            <div class="mt-3 details-title">Deadline for Signed Contract Submission:</div>
-          </v-col>
-          <v-col class="text-left mx-2">
-            <div>{{ tender.publicationDate }}</div>
-            <div class="mt-3">{{ contract.signedDeadline }}</div>
-          </v-col>
-          <v-col class="text-left mx-2">
-            <div class="details-title">Deadline for Offer Submission:</div>
-          </v-col>
-          <v-col class="text-left mx-2">
-            <div>{{ tender.offerSubmissionDeadline }}</div>
-          </v-col>
-        </v-row>
-      </v-container>
-
-      <v-container class="d-flex align-center details-container">
-        <p class="chapter-label">Documents</p>
-        <div class="horizontal-divider-details"></div>
-      </v-container>
-
-      <v-container class="details-container">
-        <v-item-group class="mx-8">
-          <v-row>
-            <FileVchip
-              :fileName="contract.fileMetadata.name"
-              :fileKey="contract.fileMetadata.awsS3fileKey"
-              @show-file="showFile"
-          ></FileVchip>
-          </v-row>
-          <div v-if="$route.params.role === USER_ROLE.CONTRACTOR" class="mt-4">
-          <v-row>
-            <FileVchip
-              :fileName="tender.awardDecision.fileMetadata.name"
-              :fileKey="tender.awardDecision.fileMetadata.awsS3fileKey"
-              @show-file="showFile"
-          ></FileVchip>
-          </v-row>
-          <v-row>
-            <FileVchip
-              :fileName="tender.rejectDecision.fileMetadata.name"
-              :fileKey="tender.rejectDecision.fileMetadata.awsS3fileKey"
-              @show-file="showFile"
-          ></FileVchip>
-            </v-row>
-          </div>
-        </v-item-group>
-      </v-container>
-        <v-row class="d-flex justify-end mt-5 mb-10">
-          <div v-if="$route.params.role === USER_ROLE.BIDDER && isNaN(Number($route.query.offer_id))">
-            <v-col md="6" style="margin-right: 20rem;">
-              <v-btn type="submit" block variant="flat" color="blue"
-                     :to="{ name: 'new-offer', query: { tender_id: $route.query.tender_id } }">
-                + Create Offer
-              </v-btn>
-            </v-col>
-         </div>
-        </v-row>
-    </v-window-item>
-    <v-window-item value="offerDescription">
-      <v-card class="mx-auto" elevation="8" max-width="1000">
-      <v-toolbar color="white" height="240" class="text-left">
-        <v-container class="pa-15">
-          <v-toolbar-title class="mb-5">Bidder</v-toolbar-title>
-          <v-row class="text-center mx-10">
-            <v-col class="text-left">
-              <div class="ma-2 details-title">Oficial Name:</div>
-              <div class="ma-2 details-title">National Registration Number: </div>
-            </v-col>
-            <v-col class="text-left">
-              <div class="ma-2">{{ offer.companyProfile.officialName }}</div>
-              <div class="ma-2">{{ offer.companyProfile.registrationNumber }}</div>
-            </v-col>
-            <v-col class="text-left">
-              <div class="ma-2 details-title">Country:</div>
-              <div class="ma-2 details-title">City / Town:</div>
-            </v-col>
-            <v-col class="text-left">
-              <div class="ma-2">{{ offer.companyProfile.country.name }}</div>
-              <div class="ma-2">{{ offer.companyProfile.city }}</div>
-            </v-col>
-          </v-row>
-        </v-container>
-      </v-toolbar>
-      </v-card>
-
-    <v-container class="d-flex align-center details-container">
-      <p class="chapter-label">Contact person</p>
-      <div class="horizontal-divider-details"></div>
-    </v-container>
-
-    <v-container class="details-container">
-      <v-row class="text-center mx-10">
-        <v-col class="text-left mx-2">
-          <div class="details-title">Name:</div>
-          <div class="mt-3 details-title">Surname:</div>
-        </v-col>
-        <v-col class="text-left mx-2">
-          <div>{{ offer.companyProfile.contactPerson.firstName }}</div>
-          <div class="mt-3">{{ offer.companyProfile.contactPerson.lastName }}</div>
-        </v-col>
-        <v-col class="text-left mx-2">
-          <div class="details-title">Phone number:</div>
-        </v-col>
-        <v-col class="text-left mx-2">
-          <div>{{ `${offer.companyProfile.country.phoneCode}${offer.companyProfile.contactPerson.phoneNumber}` }}</div>
-        </v-col>
-      </v-row>
-    </v-container>
-
-    <v-container class="d-flex align-center details-container">
-      <p class="chapter-label">Price</p>
-      <div class="horizontal-divider-details"></div>
-    </v-container>
-
-    <v-container class="details-container">
-      <v-row class="text-center mx-10">
-        <v-col class="text-left mx-2">
-          <div class="details-title">Bid Price:</div>
-        </v-col>
-        <v-col class="text-left mx-2">
-          <div>{{ offer.bidPrice }}</div>
-        </v-col>
-        <v-col class="text-left mx-2">
-          <div class="details-title">Currency:</div>
-        </v-col>
-        <v-col class="text-left mx-2">
-          <div>{{ `${offer.currency.code} | ${offer.currency.symbol}` }}</div>
-        </v-col>
-      </v-row>
-    </v-container>
-
-    <v-container class="details-container">
-      <v-item-group class="mx-8">
-        <v-row>
-          <FileVchip
-            :fileName="offer.proposition.name"
-            :fileKey="offer.proposition.awsS3fileKey"
-            @show-file="showFile"
-          ></FileVchip>
-          </v-row>
-      </v-item-group>
-    </v-container>
-    <v-container class="d-flex justify-end mt-2 mb-10">
-      <div v-if="$route.params.role === USER_ROLE.CONTRACTOR &&
-                 contract.status === CONTRACT_STATUS.DRAFT &&
-                 OFFER_STATUS[offer.status] === OFFER_STATUS.OFFER_RECEIVED">
-        <v-btn class="mx-2" type="submit" variant="outlined" color="blue"
-          @click="rejectUnsuitableOffer({ offerId: offer.id,
-                                          rejectId: tender.rejectDecision.id})">
-          Send Reject Decision
-        </v-btn>
-        <v-btn class="mx-2" type="submit" variant="flat" color="blue"
-          @click="makeAnAwardDecision({ contractId: contract.id,
-                                        offerId: offer.id,
-                                        awardId: tender.awardDecision.id})"
-          >Send Award Decision
-        </v-btn>
-      </div>
-    </v-container>
-    </v-window-item>
-
-    <FileViewerModal
-      v-model:isOpen="isOpen"
-      @update:isOpen="closeFile"
-      :fileUrl="fileUrl"
-   ></FileViewerModal>
-  </v-window>
-</template>
-
-<script>
-import { URL_REST_API } from "@/rest.api.endpoints"
-import { exceptionAlert } from "@/components/alerts";
-import { USER_ROLE, PROCEDURE, LANGUAGE, CONTRACT_TYPE, OFFER_STATUS, CONTRACT_STATUS } from "@/components/constants"
-import { fetchFromEndpoint, downloadFile, partialUpdateDocumentRecord } from "@/components/actions";
-import FileVchip from "@/components/childs/FileVchip.vue"
-import FileViewerModal from "@/components/childs/FileViewerModal.vue"
-import axios from "axios";
-import TableHeaderWithoutField from "@/components/offer/childs/TableHeaderWithoutField.vue"
-import TableBody from "@/components/offer/childs/TableBody.vue"
-
-export default {
-  components: {
-    FileVchip,
-    FileViewerModal,
-    TableHeaderWithoutField,
-    TableBody
-  },
-
-  data: () => ({
-    USER_ROLE,
-    PROCEDURE,
-    LANGUAGE,
-    CONTRACT_TYPE,
-    OFFER_STATUS,
-    CONTRACT_STATUS,
-    tender: {
-      companyProfile: {
-        country: {},
-        contactPerson: {}
-      },
-      procedure: {},
-      cpv: {},
-      awardDecision: {
-        fileMetadata: {}
-      },
-      rejectDecision: {
-        fileMetadata: {}
-      },
-    },
-    contract: {
-        currency: {},
-        contractType: {},
-        fileMetadata: {}
-      },
-    offer: {
-      hasAwardDecision: false,
-      hasRejectDecision: false
-    },
-    offers: [],
-    tab: "tenderDescription",
-    isOpen: false,
-    fileUrl: '',
-    totalPages: 1,
-    plannedPage: 1,
-    offersPerPage: 10,
-    loading: false,
-    isOffers: false,
-    exceptionAlert,
-    fetchFromEndpoint,
-    downloadFile,
-    partialUpdateDocumentRecord,
-    isTenderDecription: true
-  }),
-
-  methods: {
-    async getOffersList() {
-      try {
-        this.loading = true
-        const offersPageResponse = await axios.get(
-          `${URL_REST_API.HOST}/${URL_REST_API.OFFERS_PAGE}/${this.tender.id}?currentPage=${this.plannedPage}&totalOffers=${this.offersPerPage}`, {
-          withCredentials: true,
-          headers: {
-            'Accept': 'application/json',
-          }
-        });
-        this.totalPages = offersPageResponse.totalPages;
-        for (const offer of offersPageResponse.data.content) {
-          this.offers.push(offer);
-        }
-        if (this.offers.length == 0) {
-          this.isNoOffers = true;
-        }
-        this.plannedPage++;
-        this.loading = false;
-      } catch (error) {
-        console.log('There was an error', error)
-      }
-    },
-
-    onScroll(e) {
-      const currentPage = Math.ceil(e.target.scrollTop / 280);
-      if (currentPage === this.plannedPage && !this.loading && this.plannedPage <= this.totalPages) {
-        this.getOffersList()
-      }
-    },
-
-    async showFile(isOpen, fileKey, callback) {
-      try {
-        const response = await downloadFile(fileKey);
-        this.fileUrl = URL.createObjectURL(response.data);
-        this.isOpen = isOpen;
-      } catch (error) {
-        console.error("File download failed:", error);
-      } finally {
-        if (callback) callback();
-      }
-    },
-
-      closeFile() {
-        if (this.fileUrl) {
-          URL.revokeObjectURL(this.fileUrl);
-          this.fileUrl = null;
-        }
-        this.isOpen = false;
-      },
-
-    navigateToOffers(){
-      this.getOffersList();
-      this.tab = "offers";
-    },
-
-    async navigateToOffer(offerId) {
-      const offerResponse = await fetchFromEndpoint(`${URL_REST_API.HOST}/${URL_REST_API.OFFERS}/${offerId}`);
-      this.offer = offerResponse.data;
-      this.tab = "offerDescription"
-      console.log(this.offer)
-      if (this.$route.params.role === USER_ROLE.CONTRACTOR)
-        this.isTenderDecription = false;
-    },
-
-    async makeAnAwardDecision(awardOfferRequest) {
-      await partialUpdateDocumentRecord(awardOfferRequest, URL_REST_API.PROCUREMENTS_AWARD_OFFER);
-    },
-
-    async signContract(procurementCompletionRequest) {
-      const procurementCompletionResponse = await partialUpdateDocumentRecord(procurementCompletionRequest, URL_REST_API.PROCUREMENTS_CONTRACT_SIGN);
-      this.contract.hasSigned = procurementCompletionResponse.data.hasSigned;
-      this.offer.status = procurementCompletionResponse.data.offerStatus;
-    },
-
-    async declineContract(procurementRejectionRequest) {
-      const procurementRejectionResponse = await partialUpdateDocumentRecord(procurementRejectionRequest, URL_REST_API.PROCUREMENTS_CONTRACT_REJECT);
-    },
-
-    async rejectUnsuitableOffer(offerRejectionRequest) {
-      const offerRejectionResponse = await partialUpdateDocumentRecord(offerRejectionRequest, URL_REST_API.PROCUREMENTS_OFFER_REJECT);
-      this.offer.status = offerRejectionResponse.data.offerStatus;
-    }
-  },
-
-  async mounted() {
-    const [tenderResponse, contractResponse] = await Promise.all([
-        this.fetchFromEndpoint(`${URL_REST_API.HOST}/${URL_REST_API.TENDERS}/${this.$route.query.tender_id}`),
-        this.fetchFromEndpoint(`${URL_REST_API.HOST}/${URL_REST_API.CONTRACTS}/${this.$route.query.contract_id}`)
-      ]);
-    this.tender = tenderResponse.data;
-    this.contract = contractResponse.data;
-  }
-}
-</script>
